o/confdbstate: move mocking func to export_test
